requirements.txt
ultralytics
gradio
opencv-python
pillow
numpy
pandas
scikit-learn
matplotlib
torch
torchvision